requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-pptx==1.0.2
httpx==0.28.1
anthropic==0.42.0
lxml==5.3.0
python-multipart==0.0.20
